pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "nobox"
version = "0.1.0"
description = "JSON and YAML key-value storage utilities with CRUD operations via CLI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Jeff Mirabile", email = "jmmirabile@gmail.com"}
]
keywords = ["cli", "database", "json", "yaml", "storage", "key-value"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Utilities",
]
dependencies = [
    "confbox>=0.1.0",
    "PyYAML>=5.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/jmmirabile/nobox"
Repository = "https://github.com/jmmirabile/nobox"
Documentation = "https://github.com/jmmirabile/nobox#readme"
"Bug Tracker" = "https://github.com/jmmirabile/nobox/issues"

[project.scripts]
jsonbox = "nobox.main:main_json"
yamlbox = "nobox.main:main_yaml"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
